==> 02_data_preprocessing.ipynb <==
import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer
from scipy.stats.mstats import winsorize


## Funciones

def windsorize_columns(df, columns, limits):
    for column in columns:
        # Aplicar la winsorización a cada columna
        df[column] = winsorize(df[column], limits=limits)
    return df

## Cargo el dataset con las modificaciones iniciales

df = pd.read_csv('./data/data_preprocessing.csv', index_col=0)
df.head()

df.info()

## Primeras modificaciones

# Cambio el índice
df.set_index('ID', inplace=True)

# Cambio tipo a datetime
df['Dt_Customer'] = pd.to_datetime(df['Dt_Customer'])

# Eliminar columnas
df.drop(['Z_CostContact', 'Z_Revenue'], axis=1, inplace=True)

# Cambiar tipo a categóricas
cols_to_category = ['Education', 'Marital_Status','Complain']
df[cols_to_category] = df[cols_to_category].astype('category')

# Eliminación de YOLO y cambio de nombre a Absurd
df['Marital_Status'] = df['Marital_Status'].cat.rename_categories({'Absurd': 'Others'})
df = df[df['Marital_Status'] != 'YOLO']

df = df.drop_duplicates()


df.info()

df.loc[(df['Marital_Status'] == 'Absurd') | (df['Marital_Status']== 'YOLO' )]

## Variable target

df['Response'].value_counts(True)

Se encuentra muy desbalanceada.

## División en train y test

train_set, test_set = train_test_split(df,test_size=0.20,random_state=42,stratify=df['Response'])
print('Train_set',train_set.shape)
print('Test_set', test_set.shape)

## Imputación de nulos

Flag con los nulos

train_set['income_missing'] = np.where(train_set['Income'].isnull(),1,0)

Otras variables con correlación

corr_matrix = train_set.corr(numeric_only=True)
abs(corr_matrix['Income']).sort_values(ascending=False) > 0.5
corr_matrix[(corr_matrix['Income'] > 0.5)].index

IterativeImputer para realizar la imputación de nulos

features_iterative = ['Income', 'MntWines', 'MntMeatProducts', 'NumCatalogPurchases']
imputer_iterativo = IterativeImputer()

imputer_iterativo.fit(train_set[features_iterative])

train_set['Income'] = imputer_iterativo.transform(train_set[features_iterative])

# Aplicar al test_set
test_set['Income'] = imputer_iterativo.transform(test_set[features_iterative])

Comprobación

train_set['Income'].isna().sum()

## Nuevas variables

Imputación a train_set

# Edad tomando como referencia 2015 para poder obtener la variables
train_set['age']= 2015 - train_set['Year_Birth']

# Antigüedad como cliente tomando como referencia 2015 para poder obtener la variable
train_set['customes_seniority'] = 2015 - train_set['Dt_Customer'].dt.year

# Miembros totales en casa
dicc_status = {
    'Together': 2,  # Generalmente representa a dos personas viviendo juntas
    'Divorced': 1,  # Generalmente representa a una persona que está divorciada
    'Married': 2,   # Generalmente representa a dos personas casadas
    'Single': 1,    # Generalmente representa a una persona soltera
    'Widow': 1,     # Generalmente representa a una persona viuda
    'Alone': 1,     # Generalmente representa a una persona viviendo sola
    'Others': 1     # Puede variar, pero generalmente representa a una persona
}
train_set['Status_members'] = train_set['Marital_Status'].map(dicc_status)
train_set['Household_members'] = train_set['Status_members']+train_set['Kidhome']+train_set['Teenhome']

# Gasto total
train_set['Total_amount'] = train_set.filter(like='Mnt').sum(axis = 1)

# Nº compras totales
train_set['Total_purchase'] = train_set.filter(like='Purchases').sum(axis = 1)

# Gasto promedio
train_set['Median_amount_purchase'] = np.where(
    train_set['Total_purchase'] != 0,
    train_set['Total_amount'] / train_set['Total_purchase'],
    0)


# Nº ofertas aceptadas
train_set['Total_cmp'] = train_set.filter(like='Accepted').apply(pd.to_numeric, errors='coerce').sum(axis=1)

# % de ofertas aceptadas
train_set_numeric = train_set.filter(like='Cmp').apply(pd.to_numeric, errors='coerce')
train_set['Total_%_cmp'] = train_set['Total_cmp'] / len(train_set_numeric.columns)

# Elimino columnas intermedias
train_set = train_set.drop(columns=['Status_members'])

train_set.info()

train_set.iloc[:,25:36]

train_set['Median_amount_purchase'].describe

Imputación a test_set

# Edad tomando como referencia 2000
test_set['age'] = 2015 - test_set['Year_Birth']

# Antigüedad como cliente tomando como referencia 2020
test_set['customes_seniority'] = 2015 - test_set['Dt_Customer'].dt.year

# Miembros totales en casa
dicc_status = {
    'Together': 2,  # Generalmente representa a dos personas viviendo juntas
    'Divorced': 1,  # Generalmente representa a una persona que está divorciada
    'Married': 2,   # Generalmente representa a dos personas casadas
    'Single': 1,    # Generalmente representa a una persona soltera
    'Widow': 1,     # Generalmente representa a una persona viuda
    'Alone': 1,     # Generalmente representa a una persona viviendo sola
    'Others': 1     # Puede variar, pero generalmente representa a una persona
}
test_set['Status_members'] = test_set['Marital_Status'].map(dicc_status)
test_set['Household_members'] = test_set['Status_members'] + test_set['Kidhome'] + test_set['Teenhome']

# Gasto total
test_set['Total_amount'] = test_set.filter(like='Mnt').sum(axis=1)

# Nº compras totales
test_set['Total_purchase'] = test_set.filter(like='Purchases').sum(axis=1)

# Gasto promedio
test_set['Median_amount_purchase'] = np.where(
    test_set['Total_purchase'] != 0,
    test_set['Total_amount'] / test_set['Total_purchase'],
    0)

# Nº ofertas aceptadas
test_set['Total_cmp'] = test_set.filter(like='Accepted').apply(pd.to_numeric, errors='coerce').sum(axis=1)

# % de ofertas aceptadas
test_set_numeric = test_set.filter(like='Cmp').apply(pd.to_numeric, errors='coerce')
test_set['Total_%_cmp'] = test_set['Total_cmp'] / len(test_set_numeric.columns)

# Elimino columnas intermedias creadas para crear terceras columnas
test_set = test_set.drop(columns=['Status_members'])

test_set.info()

train_set.info()

## Tratamiento de outliers

Hay outliers en Income, con un valor de 666666 y otros 10 por encima de 120000.    
También en age con valores de más de 100 años

train_set[['Income','age']].describe().loc[['min','max']]

train_set[train_set['Income'] > 100000].shape[0]/train_set[train_set['Income'] < 100000].shape[0]

train_set[train_set['age'] > 70].shape[0]/train_set[train_set['age'] < 70].shape[0]

train_set[['Income','age']] = np.asarray(train_set[['Income','age']])
windsorize_columns(train_set,['Income','age'],[0.003,0.006])

train_set[['Income','age']].describe().loc[['min','max']]

test_set[['Income','age']] = np.asarray(test_set[['Income','age']])
windsorize_columns(test_set,['Income','age'],[0.003,0.006])

train_set.to_csv('./data/train_set.csv')
test_set.to_csv('./data/test_set.csv')
