==> 06_Modelization_smote.ipynb <==
# Modelización con aplicación de SMOTE

## Importación de librerías

# Librerías para visualización de datos
import matplotlib.pyplot as plt
import seaborn as sns

# Librerías para manipulación y análisis de datos
import numpy as np
import pandas as pd


from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import MinMaxScaler, OneHotEncoder, OrdinalEncoder
from sklearn.pipeline import Pipeline


from sklearn.feature_selection import mutual_info_classif
from sklearn.feature_selection import RFECV
from sklearn.inspection import permutation_importance
from sklearn.tree import DecisionTreeClassifier

from xgboost import XGBClassifier
from sklearn.ensemble import RandomForestClassifier
from lightgbm import LGBMClassifier
from sklearn.metrics import make_scorer, recall_score, auc

from sklearn.model_selection import cross_val_score

from sklearn.model_selection import GridSearchCV

from sklearn.metrics import confusion_matrix
from sklearn.metrics import classification_report
from sklearn.metrics import ConfusionMatrixDisplay
from sklearn.metrics import roc_curve

from imblearn.over_sampling import SMOTE
from imblearn.pipeline import Pipeline as ImbPipeline


from toolbox_DS import *
from toolbox_ML import *


import warnings
warnings.filterwarnings(action="ignore", message=r'.*Use subset.*of np.ndarray is not recommended')


## Carga de datos

Train

train_set = pd.read_csv('./data/train_set.csv')
train_set

Test

test_set = pd.read_csv('./data/test_set.csv')
test_set.head(5)

Imputo los cambios que vengo aplicando al dataset.    
Train

# Cambio el índice
train_set.set_index('ID', inplace=True)

# Cambio tipo a datetime
train_set['Dt_Customer'] = pd.to_datetime(train_set['Dt_Customer'])

# Cambiar tipo a categóricas
cols_to_category = ['Education', 'Marital_Status']
train_set[cols_to_category] = train_set[cols_to_category].astype('category')

# Elimino columna 'income_missing'
train_set = train_set.drop(columns=['income_missing', 'Year_Birth','Total_%_cmp','Dt_Customer','Median_amount_purchase'])

# Elimino el outlier de Income
train_set = train_set.loc[train_set['Income'] !=666666]

Y también al test

# Cambio el índice
test_set.set_index('ID', inplace=True)

# Cambio tipo a datetime
test_set['Dt_Customer'] = pd.to_datetime(test_set['Dt_Customer'])

# Cambiar tipo a categóricas
cols_to_category = ['Education', 'Marital_Status']
test_set[cols_to_category] = test_set[cols_to_category].astype('category')

# Elimino columna 'income_missing'
test_set = test_set.drop(columns=['Year_Birth','Total_%_cmp','Dt_Customer','Median_amount_purchase'])

# Elimino el outlier de Income
test_set = test_set.loc[test_set['Income'] !=666666]

## Transformación de variables

X_train = train_set.drop(columns='Response')
y_train = train_set['Response']

X_test = test_set.drop(columns='Response')
y_test = test_set['Response']

features_num = X_train.select_dtypes(['int','float']).columns
features_cat = X_train.select_dtypes(['object', 'category']).columns
print('features_num', features_num)
print('features_cat', features_cat)

# Definimos el OrdinalEncoder con el mapeo de 'education_ode'
ordinal_encoder = OrdinalEncoder(categories=[['Basic', '2n Cycle', 'Graduation', 'Master', 'PhD']])

categorical_features_onehot = ['Marital_Status']
categorical_features_ordinal = ['Education']
numerical_features = features_num

preprocessor = ColumnTransformer(
    transformers=[
        ('num', 'passthrough', numerical_features),
        ('cat_onehot', OneHotEncoder(), categorical_features_onehot),
        ('cat_ordinal', Pipeline([
            ('ordinal', ordinal_encoder),
            ('scaler', MinMaxScaler())
        ]), categorical_features_ordinal)
    ]
)

pipeline = Pipeline(steps=[
    ('preprocesor',preprocessor),
    ('algoritmo', RandomForestClassifier())
])

pipeline.fit(X_train,y_train)

X_train_transform = pipeline.named_steps['preprocesor'].transform(X_train)


features_transformed = pipeline.named_steps['preprocesor'].get_feature_names_out()
features_transformed

X_train_transform_df = pd.DataFrame(X_train_transform, columns=(features_transformed))
X_train_transform_df

## SMOTE

# Crear un objeto SMOTE
smote = SMOTE()

## Modelización

### Prueba de modelos con todas las variables

# ratio = float(np.sum(y_train == 0)) / np.sum(y_train == 1)
pipe = ImbPipeline(steps=[
    ('preprocesor', preprocessor),
    ('smote', smote),
    ('algoritmo', RandomForestClassifier())
])


# Definir el grid de hiperparámetros
grid = [
    {'algoritmo': [RandomForestClassifier()],
     'algoritmo__max_depth': [5, 10, 15],
     'algoritmo__n_estimators': [50, 100, 200],
     },
   
    {'algoritmo': [XGBClassifier()],
     'algoritmo__learning_rate': [0.1, 0.3, 0.5],
     'algoritmo__n_estimators': [100, 500, 1000],
     },
    
    {'algoritmo': [LGBMClassifier()],
     'algoritmo__learnin_rate': [0.1, 0.3, 0.5],
     'algoritmo__n_estimators': [100, 500, 1000],
     'algoritmo__objective': ['binary']}
]


# Realizar GridSearchCV con el scorer personalizado
grid_search = GridSearchCV(pipe, 
                           grid, 
                           cv=5, 
                           scoring=make_scorer(recall_score, pos_label=True), 
                           n_jobs=-1)


grid_search.fit(X_train, y_train)

# Obtener el mejor modelo y parámetros
best_model = grid_search.best_estimator_
print("Mejor modelo y parámetros:", grid_search.best_params_)

grid_search.best_score_

df_cv_results = pd.DataFrame(grid_search.cv_results_)
df_cv_results[['param_algoritmo','mean_test_score','std_test_score','rank_test_score']].sort_values('rank_test_score')

Haciendo un pipeline con todas las variables, el mejor sería un RandomForest con una recall de la clase positiva de 0.66

## Ajuste hiperparámetros

### RandomForest

pipe = Pipeline(steps=[
    ('preprocesador', preprocessor),
    ('algoritmo', RandomForestClassifier())
])

param = {
    'algoritmo__n_estimators': [400,450,500],
    'algoritmo__max_depth': [5, 6,7],
    'algoritmo__min_samples_split': [3,4,5],
    'algoritmo__min_samples_leaf': [8,10,12],
    'algoritmo__class_weight': ['balanced']
}

grid_rf = GridSearchCV(pipe, 
                       param_grid=param, 
                       cv=5, 
                       scoring=make_scorer(recall_score, pos_label=True), 
                       n_jobs=-1)

grid_rf.fit(X_train, y_train)

# Obtener el mejor modelo y parámetros
best_model_rf = grid_rf.best_estimator_
best_params_rf = grid_rf.best_params_
print("Mejor modelo y parámetros:", best_params_rf)

# Imprimir el mejor resultado del recall para la clase positiva
best_recall_rf = grid_rf.best_score_
print("Mejor resultado del recall para la clase positiva:", best_recall_rf)


# Acceder al modelo dentro del Pipeline
random_forest_model = best_model_rf.named_steps['algoritmo']

# Obtener las importancias de las características
feature_importances = random_forest_model.feature_importances_

if 'preprocesador' in best_model_rf.named_steps:
    preprocessor = best_model_rf.named_steps['preprocesador']
    feature_names = preprocessor.get_feature_names_out()

    importances_series = pd.Series(feature_importances, index=feature_names)
    print(importances_series.sort_values(ascending=False))

### LigthGBM

pipe = Pipeline(steps=[
    ('preprocesador', preprocessor),
    ('algoritmo', LGBMClassifier())
])

param = {
    'algoritmo__max_depth': [4,5, 6],
    'algoritmo__min_child_samples': [2,4,8],
    'algoritmo__min_child_weight': [4, 8,10],
    'algoritmo__learning_rate': [0.09, 0.1, 0.2],
    'algoritmo__n_estimators': [100, 200, 300],
    'algoritmo__is_unbalance' : [True],
    'algoritmo__objective': ['binary'],
}

grid_lgb = GridSearchCV(pipe, 
                       param_grid=param, 
                       cv=5, 
                       scoring=make_scorer(recall_score, pos_label=True), 
                       n_jobs=-1,
                       verbose= 1)

grid_lgb.fit(X_train, y_train)

# Obtener el mejor modelo y parámetros
best_model_lgb = grid_lgb.best_estimator_
best_params_lgb = grid_lgb.best_params_
print("Mejor modelo y parámetros:", best_params_lgb)

# Imprimir el mejor resultado del recall para la clase positiva
best_recall_lgb = grid_lgb.best_score_
print("Mejor resultado del recall para la clase positiva:", best_recall_lgb)

# Acceder al modelo dentro del Pipeline
lgb_model = best_model_lgb.named_steps['algoritmo']

# Obtener las importancias de las características
feature_importances = lgb_model.feature_importances_

if 'preprocesador' in best_model_lgb.named_steps:
    preprocessor = best_model_lgb.named_steps['preprocesador']
    feature_names = preprocessor.get_feature_names_out()

    importances_series = pd.Series(feature_importances, index=feature_names)
    print(importances_series.sort_values(ascending=False))

### XGBoost

pipe = Pipeline(steps=[
    ('preprocesador', preprocessor),
    ('algoritmo', XGBClassifier())
])

param = {
    'algoritmo__n_estimators': [300, 400, 450],
    'algoritmo__max_depth': [4, 5, 6],
    'algoritmo__min_child_weight': [6, 8, 10],  # Equivalente a min_samples_leaf en RandomForest
    'algoritmo__gamma': [0, 0.1, 0.2],  # Parámetro para controlar la regularización
    'algoritmo__subsample': [0.7, 0.8, 0.9],  # Porcentaje de muestras usadas para cada árbol
    'algoritmo__colsample_bytree': [0.7, 0.8, 0.9],  # Porcentaje de características usadas para cada árbol
    
}


grid_xgb = GridSearchCV(pipe, 
                       param_grid=param, 
                       cv=5, 
                       scoring=make_scorer(recall_score, pos_label=True), 
                       n_jobs=-1,
                       verbose= 3)

grid_xgb.fit(X_train, y_train)

# Obtener el mejor modelo y parámetros
best_model_xgb = grid_xgb.best_estimator_
best_params_xgb = grid_xgb.best_params_
print("Mejor modelo y parámetros:", best_params_xgb)

# Imprimir el mejor resultado del recall para la clase positiva
best_recall_xgb = grid_xgb.best_score_
print("Mejor resultado del recall para la clase positiva:", best_recall_xgb)

# Acceder al modelo dentro del Pipeline
xgb_model = best_model_xgb.named_steps['algoritmo']

# Obtener las importancias de las características
feature_importances = xgb_model.feature_importances_

if 'preprocesador' in best_model_xgb.named_steps:
    preprocessor = best_model_xgb.named_steps['preprocesador']
    feature_names = preprocessor.get_feature_names_out()

    importances_series = pd.Series(feature_importances, index=feature_names)
    print(importances_series.sort_values(ascending=False))

## Evaluación Modelos

#### Classification Report

# RandomForest
print('RandomForest')
y_pred_rf = grid_rf.predict(X_test)

print(classification_report(y_test,y_pred_rf))
print()

# LigthGBM
print('LightGBM')
y_pred_lgb = grid_lgb.predict(X_test)
print(classification_report(y_test,y_pred_lgb))
print()

# XGBoost
print('LightGBM')
y_pred_xgb = grid_xgb.predict(X_test)
print(classification_report(y_test,y_pred_xgb))
print()



#### Matrix de confusión

# RandomForest
print('RandomForest')
cm_rf = confusion_matrix(y_test, y_pred_rf)
disp_rf = ConfusionMatrixDisplay(confusion_matrix=cm_rf)
disp_rf.plot()
print()

# LightGBM
print('LightGBM')
cm_lgb = confusion_matrix(y_test, y_pred_lgb)
disp_lgb = ConfusionMatrixDisplay(confusion_matrix=cm_lgb)
disp_lgb.plot()
print()

# XGBoost
print('XGBoost')
cm_xgb = confusion_matrix(y_test, y_pred_xgb)
disp_xgb = ConfusionMatrixDisplay(confusion_matrix=cm_xgb)
disp_xgb.plot()
print()
